Adapted CudaAPI to the new DLL-interop API.
