+++ Src/ILGPU/Runtime/Cuda/CudaStream.cs
@@ -9,10 +9,10 @@
 // Source License. See LICENSE.txt for details
 // ---------------------------------------------------------------------------------------
 
-using ILGPU.Runtime.Cuda.API;
 using ILGPU.Util;
 using System;
 using System.Diagnostics.CodeAnalysis;
+using static ILGPU.Runtime.Cuda.CudaAPI;
 
 namespace ILGPU.Runtime.Cuda
 {
@@ -55,7 +55,7 @@ namespace ILGPU.Runtime.Cuda
             : base(accelerator)
         {
             CudaException.ThrowIfFailed(
-                CudaAPI.Current.CreateStream(
+                CurrentAPI.CreateStream(
                     out streamPtr,
                     flag));
             responsibleForHandle = true;
@@ -80,7 +80,7 @@ namespace ILGPU.Runtime.Cuda
             var binding = Accelerator.BindScoped();
 
             CudaException.ThrowIfFailed(
-                CudaAPI.Current.SynchronizeStream(streamPtr));
+                CurrentAPI.SynchronizeStream(streamPtr));
 
             binding.Recover();
         }
@@ -95,7 +95,7 @@ namespace ILGPU.Runtime.Cuda
             if (responsibleForHandle && streamPtr != IntPtr.Zero)
             {
                 CudaException.ThrowIfFailed(
-                    CudaAPI.Current.DestroyStream(streamPtr));
+                    CurrentAPI.DestroyStream(streamPtr));
                 streamPtr = IntPtr.Zero;
             }
             base.Dispose(disposing);
